Move mapif functions from int_elemental.c to mapif.c
